+++ Engine/Core/Core.vcxproj.filters
@@ -20,7 +20,6 @@
     <ClCompile Include="src\yougine\Editor\SceneWindow.cpp" />
     <ClCompile Include="src\yougine\components\RenderComponent.cpp" />
     <ClCompile Include="src\yougine\Editor\SelectionInfo.cpp" />
-    <ClCompile Include="src\yougine\Layer.cpp" />
     <ClCompile Include="src\yougine\managers\GameManager.cpp" />
     <ClCompile Include="src\yougine\managers\RenderManager.cpp" />
     <ClCompile Include="src\yougine\components\Component.cpp" />
@@ -188,7 +187,6 @@
     <ClInclude Include="src\yougine\Editor\SceneWindow.h" />
     <ClInclude Include="src\yougine\components\RenderComponent.h" />
     <ClInclude Include="src\yougine\Editor\SelectionInfo.h" />
-    <ClInclude Include="src\yougine\Layer.h" />
     <ClInclude Include="src\yougine\managers\GameManager.h" />
     <ClInclude Include="src\yougine\managers\IManager.h" />
     <ClInclude Include="src\yougine\managers\RenderManager.h" />
@@ -282,7 +280,6 @@
     <ClInclude Include="src\yougine\Editor\HierarchyWindow.h" />
     <ClInclude Include="src\yougine\InputManager.h" />
     <ClInclude Include="src\yougine\managers\ComponentList.h" />
-    <ClInclude Include="src\yougine\managers\ComponentName.h" />
     <ClInclude Include="src\yougine\managers\CustomScriptManager.h" />
     <ClInclude Include="src\yougine\GameObject.h" />
     <ClInclude Include="src\yougine\Scene.h" />
