+++ Engine/Core/src/yougine/Scene.h
@@ -2,7 +2,7 @@
 #include "GameObject.h"
 #include "managers/ComponentList.h"
 #include <list>
-#include "Layer.h"
+#include "UserShare/Layer.h"
 
 namespace yougine
 {
