+++ Engine/CoreExeBuildProject/CoreExeBuildProject.vcxproj
@@ -116,7 +116,7 @@
       <SDLCheck>true</SDLCheck>
       <PreprocessorDefinitions>_DEBUG;_CONSOLE;%(PreprocessorDefinitions);GLEW_STATIC</PreprocessorDefinitions>
       <ConformanceMode>true</ConformanceMode>
-      <AdditionalIncludeDirectories>$(SolutionDir)Core\src\yougine;$(SolutionDir)Core\lib\glfw\include;$(SolutionDir)Core\lib\glew\include;$(SolutionDir)Core\lib;$(SolutionDir)UserEngineCommon</AdditionalIncludeDirectories>
+      <AdditionalIncludeDirectories>$(SolutionDir)Core\src\yougine;$(SolutionDir)Core\lib\glfw\include;$(SolutionDir)Core\lib\glew\include;$(SolutionDir)Core\lib;$(SolutionDir)UserEngineCommon;$(SolutionDir)UserEngineCommon\include</AdditionalIncludeDirectories>
       <LanguageStandard>stdcpp17</LanguageStandard>
     </ClCompile>
     <Link>
@@ -134,7 +134,7 @@
       <SDLCheck>true</SDLCheck>
       <PreprocessorDefinitions>NDEBUG;_CONSOLE;%(PreprocessorDefinitions);GLEW_STATIC</PreprocessorDefinitions>
       <ConformanceMode>true</ConformanceMode>
-      <AdditionalIncludeDirectories>$(SolutionDir)Core\src\yougine;$(SolutionDir)Core\lib\glfw\include;$(SolutionDir)Core\lib\glew\include;$(SolutionDir)Core\lib;$(SolutionDir)UserEngineCommon</AdditionalIncludeDirectories>
+      <AdditionalIncludeDirectories>$(SolutionDir)Core\src\yougine;$(SolutionDir)Core\lib\glfw\include;$(SolutionDir)Core\lib\glew\include;$(SolutionDir)Core\lib;$(SolutionDir)UserEngineCommon;$(SolutionDir)UserEngineCommon\include</AdditionalIncludeDirectories>
     </ClCompile>
     <Link>
       <SubSystem>Console</SubSystem>
