+++ Engine/UserEngineCommon/UserEngineCommon.vcxproj.filters
@@ -24,9 +24,15 @@
     <ClInclude Include="include\UserShare\utilitys\YougineMath.h">
       <Filter>ヘッダー ファイル</Filter>
     </ClInclude>
+    <ClInclude Include="include\UserShare\Layer.h">
+      <Filter>ヘッダー ファイル</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="CPP\UserShare\utilitys\Split.cpp">
+    <ClCompile Include="include\UserShare\utilitys\Split.cpp">
+      <Filter>ソース ファイル</Filter>
+    </ClCompile>
+    <ClCompile Include="include\UserShare\Layer.cpp">
       <Filter>ソース ファイル</Filter>
     </ClCompile>
   </ItemGroup>
